split the ReqClient test in client and server parts
